refactor(hopr-strategy): split channel_lifecycle into mod/strategy/pipeline/events
Moves the monolithic channel_lifecycle module into four logically-isolated
files: mod.rs (struct defs + tests), strategy.rs (builder + run loop),
pipeline.rs (5-pass decision logic + action dispatchers), events.rs
(fast-lane chain event handlers). All 10 tests pass.
