+++ impls/strategy/src/channel_lifecycle/strategy.rs
@@ -1,0 +1,169 @@
+//! Builder, `Display`, and `Strategy` trait implementation.
+
+use std::{
+    fmt::{Display, Formatter},
+    sync::Arc,
+    time::Duration,
+};
+
+use async_trait::async_trait;
+use dashmap::{DashMap, DashSet};
+use futures::StreamExt as _;
+use hopr_lib::api::{
+    chain::{
+        ChainEvent, ChainReadAccountOperations, ChainReadChannelOperations, ChainReadSafeOperations, ChainValues,
+        ChainWriteChannelOperations,
+    },
+    node::{
+        ActionableEvent, ActionableEventDiscriminant, ActionableEventSource, HasChainApi, HasGraphView, HasNetworkView,
+    },
+};
+
+use crate::{errors::StrategyError, strategy::Strategy as StrategyTrait};
+
+use super::{ChannelLifecycleConfig, ChannelLifecycleStrategyInner};
+
+/// Builder for [`ChannelLifecycleStrategy`].
+///
+/// Cadence (`tick_interval`, `jitter`) lives in the config struct — operators
+/// can tune it without touching wiring code.
+///
+/// Call [`new`](ChannelLifecycleStrategy::new) with the strategy configuration,
+/// then [`build`](ChannelLifecycleStrategy::build) to wire in a node and obtain
+/// a runnable `Box<dyn Strategy + Send>`.
+pub struct ChannelLifecycleStrategy {
+    cfg: ChannelLifecycleConfig,
+}
+
+impl ChannelLifecycleStrategy {
+    /// Create a new builder with the given configuration.
+    pub fn new(cfg: ChannelLifecycleConfig) -> Self {
+        Self { cfg }
+    }
+
+    /// Wire in a node and return a running-ready strategy.
+    ///
+    /// The generic `N` is erased at construction time; the returned
+    /// `Box<dyn Strategy + Send>` can be held and spawned without knowledge
+    /// of the concrete node type.
+    pub fn build<N>(self, node: Arc<N>) -> Box<dyn StrategyTrait + Send>
+    where
+        N: HasChainApi + HasNetworkView + HasGraphView + ActionableEventSource + Send + Sync + 'static,
+        N::ChainApi: ChainReadChannelOperations
+            + ChainReadSafeOperations
+            + ChainReadAccountOperations
+            + ChainValues
+            + ChainWriteChannelOperations
+            + Clone
+            + Send
+            + Sync
+            + 'static,
+    {
+        Box::new(ChannelLifecycleStrategyInner {
+            cfg: self.cfg,
+            node,
+            open_in_flight: Arc::new(DashSet::new()),
+            fund_in_flight: Arc::new(DashSet::new()),
+            close_in_flight: Arc::new(DashSet::new()),
+            finalize_in_flight: Arc::new(DashSet::new()),
+            cooldown: Arc::new(DashMap::new()),
+            start_epoch: std::time::Instant::now(),
+            last_observed: Arc::new(DashMap::new()),
+            peer_ticket_activity: Arc::new(DashMap::new()),
+        })
+    }
+}
+
+impl<N> Display for ChannelLifecycleStrategyInner<N> {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(f, "channel_lifecycle")
+    }
+}
+
+#[async_trait]
+impl<N> StrategyTrait for ChannelLifecycleStrategyInner<N>
+where
+    N: HasChainApi + HasNetworkView + HasGraphView + ActionableEventSource + Send + Sync + 'static,
+    N::ChainApi: ChainReadChannelOperations
+        + ChainReadSafeOperations
+        + ChainReadAccountOperations
+        + ChainValues
+        + ChainWriteChannelOperations
+        + Clone
+        + Send
+        + Sync
+        + 'static,
+{
+    async fn run(&mut self) -> crate::errors::Result<()> {
+        self.run_pipeline().await;
+
+        let me = *self.node.chain_api().me();
+
+        // Derive a fixed per-run jitter offset from system-time nanoseconds so
+        // nodes restarted simultaneously spread out their ticks.
+        let jitter_ns = self.cfg.jitter.as_nanos() as u64;
+        let jitter_offset = if jitter_ns > 0 {
+            Duration::from_nanos(
+                std::time::SystemTime::now()
+                    .duration_since(std::time::UNIX_EPOCH)
+                    .unwrap_or_default()
+                    .subsec_nanos() as u64
+                    % jitter_ns,
+            )
+        } else {
+            Duration::ZERO
+        };
+        let effective_interval = self.cfg.tick_interval + jitter_offset;
+
+        let tick_stream = futures_time::stream::interval(effective_interval.into()).map(|_| LoopEvent::Tick);
+
+        let event_stream = self
+            .node
+            .subscribe_to_actionable_events(Some(&[ActionableEventDiscriminant::Chain]))
+            .map_err(|e| StrategyError::Other(anyhow::anyhow!(e)))?
+            .filter_map(|ev| {
+                futures::future::ready(match ev {
+                    ActionableEvent::Chain(e) => Some(LoopEvent::Chain(Box::new(e))),
+                    _ => None,
+                })
+            });
+
+        let mut driver = futures_concurrency::stream::Merge::merge((tick_stream, event_stream));
+
+        while let Some(evt) = driver.next().await {
+            match evt {
+                LoopEvent::Tick => {
+                    self.run_pipeline().await;
+                }
+                LoopEvent::Chain(e) => match *e {
+                    ChainEvent::ChannelBalanceDecreased(ch, _) => {
+                        self.on_balance_decreased(ch, me).await;
+                    }
+                    ChainEvent::ChannelBalanceIncreased(ch, _) => {
+                        self.on_balance_increased(ch);
+                    }
+                    ChainEvent::ChannelOpened(ch) => {
+                        self.on_channel_opened(ch);
+                    }
+                    ChainEvent::ChannelClosureInitiated(ch) => {
+                        self.on_channel_closure_initiated(ch);
+                    }
+                    ChainEvent::ChannelClosed(ch) => {
+                        self.on_channel_closed(ch);
+                    }
+                    ChainEvent::TicketRedeemed(ch, _) => {
+                        self.on_ticket_redeemed(ch);
+                    }
+                    _ => {}
+                },
+            }
+        }
+
+        Ok(())
+    }
+}
+
+enum LoopEvent {
+    Tick,
+    Chain(Box<ChainEvent>),
+}
